Refactor a bit, add loot functions
